fix travis again

+++ c_plus_plus/src/model/classifier/classifier.cpp
@@ -1,3 +1,5 @@
+#include <cmath>
+
 #include "classifier.h"
 
 namespace Model
@@ -13,8 +15,8 @@ std::vector<int> Classifier::classify(std::vector<SourcePosition> audioPositions
         {
             for(size_t j = 0; j < imagePositions.size(); j++)
             {
-                if(std::abs(audioPositions[i].azimuth - imagePositions[j].azimuth) > static_cast<float>(rangeThreshold)
-                        || std::abs(audioPositions[i].elevation - imagePositions[j].elevation) > static_cast<float>(rangeThreshold))
+                if(std::abs(audioPositions[i].azimuth - imagePositions[j].azimuth) > rangeThreshold
+                        || std::abs(audioPositions[i].elevation - imagePositions[j].elevation) > rangeThreshold)
                 {
                     sourcesToSuppress.push_back(static_cast<int>(i));
                 }
